refactor: refactored navtab component
- created BusinessNavTab component to replace NavTabs component.
- modified it to make it DRY.
