+++ src/components/business/Business.jsx
@@ -1,11 +1,12 @@
 import BusinessHeader from './partials/BusinessHeader'
 import BusinessCards from './partials/BusinessCards'
-
+import BusinessNavTab from './partials/BusinessNavTab'
 const Business = () => {
   return (
     <>
       <BusinessHeader />
       <BusinessCards />
+      <BusinessNavTab />
     </>
   )
 }
