+++ src/pages/Home.jsx
@@ -2,7 +2,7 @@ import About from '../components/about/About'
 import Footer from '../components/footer/Footer'
 import Header from '../components/header/Header'
 import Hero from '../components/hero/Hero'
-import NavTabs from '../components/navtabs/NavTabs'
+// import NavTabs from '../components/navtabs/NavTabs'
 import Business from '../components/business/Business'
 
 const Home = () => {
@@ -11,7 +11,7 @@ const Home = () => {
       <Header />
       <Hero />
       <Business />
-      <NavTabs />
+      {/* <NavTabs /> */}
       <About />
       <Footer />
     </>
