+++ src/styles/partials/_header.scss
@@ -73,16 +73,6 @@ nav {
   transition: all 0.5s ease-in-out;
 }
 
-// .menu-list li:last-child a {
-//   font-weight: 300;
-//   padding: 5px 10px;
-//   color: #fff;
-//   background-color: #01aacd;
-//   transition: all 0.5s ease-in-out;
-//   margin-left: 20px;
-//   border-radius: 2px;
-// }
-
 @media screen and (max-width: 800px) {
   .menu-list li a {
     font-size: 14px;
@@ -116,7 +106,7 @@ nav {
   }
 
   .menu-list.close {
-    right: -110%;
+    right: -120%;
     transition: all 0.5s ease-in-out;
   }
 }
